requirements.txt
acres==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
etelemetry==0.3.1
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
fitz==0.0.1.dev2
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-nomic==0.1.4
langchain-ollama==0.2.2
langchain-postgres==0.0.12
langchain-text-splitters==0.3.5
langsmith==0.2.10
loguru==0.7.3
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nomic==3.3.4
numpy==1.26.4
ollama==0.4.5
orjson==3.10.14
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pgvector==0.2.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
prov==2.0.1
psutil==6.1.1
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.1.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
yarl==1.18.3
